[native] Create MessageListContextContainer
Summary: To render NodeHeightMeasurer it has to be a descendant of MessageListContext. The new chat context will be rendered outside of any MessageListContexts in a class component. To be able to easily wrap something with MessageListContext, especially when threadID might change, I created a wrapper component which uses the provider.

Test Plan: Tested with other diffs

Reviewers: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1402