[native] Implement chat container
Summary:
Most of the logic is copied from `MessageListContainer`. I'll mark the differences with comments.
The most interesting part is a queue with `pendingMeasurements`. So every time a client wants to measure messages, we add a task to a queue or set it as a `currentMeasurement`. This task consists of a promise which is returned to the caller. Later, in `allHeightsMeasured` we resolve this promise and that results in client receiving the result.

Test Plan: Tested with other diffs

Reviewers: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1403