[native] Use `measuredHeights` to initialize new ChatItemHeightMeasurers
Summary: Check if the client has a measurement task. If that's the case, it means that `ChatItemHeightMeasurer` was already created and doesn't need to be initialized. When we're about to create a new measurer, we need to check if there's an existing measurer for the same thread and use saved value to initialize the new one.

Test Plan: Testing was quite complicated. I modified text tooltip button to measure messages for a pending sidebar. Then I opened the sidebar and verified that new measurer was not creating dummies for the messages. It means that the old result was used instead of measuring again.

Reviewers: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1507