[native] Delete onLongPress from MultimediaMessageMultimedia
Summary: This gesture will be handled by the whole message. We're going to change nav params and modal rendering code, so it is a good idea to disable the gesture so that app don't crash when a user long presses an image.

Test Plan: Long pressing the image should not show the modal.

Reviewers: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1630