[native] Provide onPressMultimedia callback to inner multimedia message
Summary: When a multimedia is rendered by a tooltip, we don't want to perform the measurement. To do that, we can introduce a callback with measure handler code and decide if we want to measure based on the presence of the callback.

Test Plan: Click an image and check if image modal is displayed correctly.

Reviewers: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1636